fix accidentally moved contents in curve
